# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "该系统用于管理自习室的预约、签到、退座等功能。",
        "title": "自习室管理系统 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support"
        },
        "version": "1.0"
    },
    "host": "localhost:7077",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/seat": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seat"
                ],
                "summary": "获取座位列表（支持分页和条件过滤）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "自习室ID",
                        "name": "room_id",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否有插座",
                        "name": "has_socket",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否可用",
                        "name": "is_available",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码，从1开始",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.Seat"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seat"
                ],
                "summary": "创建单个座位",
                "parameters": [
                    {
                        "description": "座位参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.createSeatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.Seat"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/admin/seat/batch": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seat"
                ],
                "summary": "批量创建座位（用于初始化）",
                "parameters": [
                    {
                        "description": "座位数组及所属自习室",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.createSeatsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.Seat"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seat"
                ],
                "summary": "批量更新座位信息（如状态/插座）",
                "parameters": [
                    {
                        "description": "座位批量更新参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.updateSeatsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.Seat"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/admin/seat/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seat"
                ],
                "summary": "获取座位详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "座位ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.Seat"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seat"
                ],
                "summary": "删除座位",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "座位ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seat"
                ],
                "summary": "更新座位信息（不可更改RoomID）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "座位ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "可修改字段",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.updateSeatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.Seat"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/v1/token/refresh": {
            "post": {
                "description": "使用refresh token刷新access token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "刷新访问令牌",
                "parameters": [
                    {
                        "description": "Refresh Token Request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.refreshAccessTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.refreshAccessTokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.createSeatRequest": {
            "type": "object",
            "required": [
                "has_socket",
                "room_id",
                "seat_num"
            ],
            "properties": {
                "has_socket": {
                    "type": "boolean"
                },
                "is_available": {
                    "type": "boolean"
                },
                "room_id": {
                    "type": "integer",
                    "minimum": 1
                },
                "seat_num": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
        "api.createSeatsRequest": {
            "type": "object",
            "required": [
                "room_id",
                "seats"
            ],
            "properties": {
                "room_id": {
                    "type": "integer",
                    "minimum": 1
                },
                "seats": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "object",
                        "required": [
                            "has_socket",
                            "is_available",
                            "seat_num"
                        ],
                        "properties": {
                            "has_socket": {
                                "type": "boolean"
                            },
                            "is_available": {
                                "type": "boolean"
                            },
                            "seat_num": {
                                "type": "string",
                                "minLength": 1
                            }
                        }
                    }
                }
            }
        },
...
        "api.refreshAccessTokenRequest": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "api.refreshAccessTokenResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "access_token_expires_at": {
                    "type": "string"
                }
            }
        },
...
        "api.updateSeatRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "has_socket": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer",
                    "minimum": 1
                },
                "is_available": {
                    "type": "boolean"
                },
                "seat_num": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
        "api.updateSeatsRequest": {
            "type": "object",
            "required": [
                "seats"
            ],
            "properties": {
                "seats": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "object",
                        "required": [
                            "has_socket",
                            "ids",
                            "is_available",
                            "seat_num"
                        ],
                        "properties": {
                            "has_socket": {
                                "type": "boolean"
                            },
                            "ids": {
                                "type": "integer",
                                "minimum": 1
                            },
                            "is_available": {
                                "type": "boolean"
                            },
                            "seat_num": {
                                "type": "string",
                                "minLength": 1
                            }
                        }
                    }
                }
            }
        },
...
        "db.Seat": {
            "type": "object",
            "properties": {
                "has_socket": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "is_available": {
                    "type": "boolean"
                },
                "number": {
                    "type": "string"
                },
                "room_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}